+++ src/xsd_training/models/__init__.py
@@ -1,0 +1,4 @@
+from .group import *
+from .lesson import *
+from .qualification import *
+from .sdc import *